Rename weather forecast to weather forecast api
* Update pages that use variables from the file
* Move file from api folder to weather component folder
