+++ client/src/components/Weather/Weather.jsx
@@ -1,6 +1,6 @@
 import { useEffect } from "react";
 import { useState } from "react";
-import { getBeachWeatherData } from "../../api/WeatherForecast";
+import { getBeachWeatherData } from "./WeatherForecastApi";
 import { WeatherCard } from "./WeatherCard";
 
 export const Weather = ({ beach }) => {
