+++ client/src/components/Weather/WeatherForecastApi.js
@@ -1,4 +1,4 @@
-import { weatherAPI_KEY } from "../weatherApiKey";
+import { weatherAPI_KEY } from "../../weatherApiKey";
 const BASE_URL =
   "https://weather.visualcrossing.com/VisualCrossingWebServices/rest/services/weatherdata/forecast";
